+++ pkg/files/files.go
@@ -1,6 +1,5 @@
-// Package util provides utility functions for file discovery and related
-// helpers used across the project.
-package util
+// Package files provides utility functions for file discovery and related helpers.
+package files
 
 import (
 	"context"
@@ -10,11 +9,11 @@ import (
 	"path/filepath"
 )
 
-// FindFiles searches recursively from the current directory for files with the
+// RecursiveFind searches recursively from the current directory for files with the
 // given name. It returns a slice of matching file paths or an error if
 // directory traversal fails. Logging is performed for each found file using
 // slog with the provided context.
-func FindFiles(ctx context.Context, name string) ([]string, error) {
+func RecursiveFind(ctx context.Context, name string) ([]string, error) {
 	var files []string
 
 	err := filepath.WalkDir(".", func(path string, d os.DirEntry, err error) error {
